requirements.txt

python-dotenv
pandas
requests
llama-index
llama-index-llms-groq
llama-index-embeddings-nvidia
pyarrow
llama_index.embeddings.huggingface
llama-index-readers-file
arxiv
llama-index-tools-tavily-research
crewai
crewai-tools
